requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-dotenv==1.0.0
google-generativeai==0.8.3
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
aiofiles==23.2.1
faiss-cpu==1.7.4
